# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AddBlockResponse:
    description: Ответ на добавление блока
    properties:
      blockId:
        example: block_123
        type: string
      message:
        example: Block added successfully
        type: string
    type: object
...
  handlers.CreateMaterialResponse:
    description: Ответ на создание материала
    properties:
      editorUrl:
        example: /editor/1
        type: string
      material:
        $ref: '#/definitions/models.Material'
      message:
        example: Material created successfully
        type: string
    type: object
...
  handlers.ForbiddenErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Access forbidden
        type: string
    type: object
  handlers.InternalErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Internal server error
        type: string
    type: object
...
  handlers.InvalidIDErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid ID
        type: string
    type: object
  handlers.InvalidParametersErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Invalid request parameters
        type: string
    type: object
...
  handlers.MaterialNotFoundErrorResponse:
    description: Стандартный ответ с ошибкой
    properties:
      error:
        example: Material is not found
        type: string
    type: object
...
  handlers.PublishMaterialResponse:
    description: Ответ на публикацию материала
    properties:
      material:
        $ref: '#/definitions/models.Material'
      message:
        example: Material published successfully
        type: string
      shareUrl:
        example: https://paydeya.com/share/abc123
        type: string
    type: object
...
  handlers.ReorderBlocksRequest:
    description: Запрос на изменение порядка блоков
    properties:
      blocks:
        example:
        - block_1
        - block_2
        - block_3
        items:
          type: string
        type: array
    required:
    - blocks
    type: object
  handlers.ReorderBlocksResponse:
    description: Ответ на изменение порядка блоков
    properties:
      message:
        example: Blocks reordered successfully
        type: string
      newOrder:
        example:
        - block_1
        - block_2
        - block_3
        items:
          type: string
        type: array
    type: object
...
  handlers.SuccessResponse:
    description: Стандартный успешный ответ
    properties:
      message:
        example: Operation completed successfully
        type: string
    type: object
...
  models.AnimationStep:
    description: Шаг анимации
    properties:
      action:
        description: show, hide, highlight
        example: show
        type: string
      element:
        example: element_1
        type: string
      style:
        additionalProperties: true
        type: object
    type: object
...
  models.Block:
    description: Блок контента в материале
    properties:
      animation:
        $ref: '#/definitions/models.BlockAnimation'
      content:
        additionalProperties: true
        type: object
      id:
        example: block_123
        type: string
      position:
        example: 1
        type: integer
      styles:
        additionalProperties: true
        type: object
      type:
        description: text, image, video, formula, quiz
        example: text
        type: string
    type: object
  models.BlockAnimation:
    description: Анимация блока
    properties:
      delay:
        example: 1000
        type: integer
      steps:
        items:
          $ref: '#/definitions/models.AnimationStep'
        type: array
      trigger:
        description: click, auto
        example: click
        type: string
    type: object
...
  models.CreateMaterialRequest:
    description: Запрос на создание материала
    properties:
      subject:
        example: math
        type: string
      title:
        example: Новый материал
        type: string
    required:
    - subject
    - title
    type: object
...
  models.Material:
    description: Учебный материал
    properties:
      access:
        description: open, link
        example: open
        type: string
      authorId:
        example: 123
        type: integer
      authorName:
        example: Иван Иванов
        type: string
      blocks:
        items:
          $ref: '#/definitions/models.Block'
        type: array
      createdAt:
        example: "2023-01-15T10:30:00Z"
        type: string
      id:
        example: 1
        type: integer
      shareUrl:
        example: https://paydeya.com/share/abc123
        type: string
      status:
        description: draft, published, archived
        example: published
        type: string
      subject:
        example: math
        type: string
      title:
        example: Основы алгебры
        type: string
      updatedAt:
        example: "2023-01-15T10:30:00Z"
        type: string
    type: object
...
  models.PublishMaterialRequest:
    description: Запрос на публикацию материала
    properties:
      access:
        description: open, link
        example: open
        type: string
      visibility:
        description: draft, published, archived
        example: published
        type: string
    type: object
...
  models.UpdateMaterialRequest:
    description: Запрос на обновление материала
    properties:
      blocks:
        items:
          $ref: '#/definitions/models.Block'
        type: array
      title:
        example: Обновленное название
        type: string
    type: object
...
  models.UserMaterialsResponse:
    description: Ответ со списком материалов пользователя
    properties:
      materials:
        items:
          $ref: '#/definitions/models.Material'
        type: array
      status:
        example: draft
        type: string
      total:
        example: 5
        type: integer
      userID:
        example: 123
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для образовательной платформы Пайдея
  title: Paydeya Education Platform API
  version: "1.0"
paths:
...
  /materials:
    post:
      consumes:
      - application/json
      description: Создает новый учебный материал
      parameters:
      - description: Данные материала
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.CreateMaterialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Материал создан
          schema:
            $ref: '#/definitions/handlers.CreateMaterialResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Создать материал
      tags:
      - materials
  /materials/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает материал по ID
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Материал
          schema:
            $ref: '#/definitions/models.Material'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "401":
          description: Неверный ID
          schema:
            $ref: '#/definitions/handlers.InvalidIDErrorResponse'
        "404":
          description: Материал не найден
          schema:
            $ref: '#/definitions/handlers.MaterialNotFoundErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить материал
      tags:
      - materials
    put:
      consumes:
      - application/json
      description: Обновляет материал (только для автора)
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Материал обновлен
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "403":
          description: Доступ запрещен
          schema:
            $ref: '#/definitions/handlers.ForbiddenErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить материал
      tags:
      - materials
  /materials/{id}/blocks:
    post:
      consumes:
      - application/json
      description: Добавляет блок к материалу
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Данные блока
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Block'
      produces:
      - application/json
      responses:
        "200":
          description: Блок добавлен
          schema:
            $ref: '#/definitions/handlers.AddBlockResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Добавить блок
      tags:
      - materials
  /materials/{id}/blocks/{blockId}:
    delete:
      consumes:
      - application/json
      description: Удаляет блок из материала
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: ID блока
        in: path
        name: blockId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Блок удален
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить блок
      tags:
      - materials
    put:
      consumes:
      - application/json
      description: Обновляет блок материала
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: ID блока
        in: path
        name: blockId
        required: true
        type: string
      - description: Данные блока
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.Block'
      produces:
      - application/json
      responses:
        "200":
          description: Блок обновлен
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить блок
      tags:
      - materials
  /materials/{id}/blocks/reorder:
    post:
      consumes:
      - application/json
      description: Изменяет порядок блоков в материале
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Новый порядок блоков
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ReorderBlocksRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Порядок изменен
          schema:
            $ref: '#/definitions/handlers.ReorderBlocksResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Изменить порядок блоков
      tags:
      - materials
  /materials/{id}/publish:
    post:
      consumes:
      - application/json
      description: Публикует материал с указанными настройками видимости
      parameters:
      - description: ID материала
        in: path
        name: id
        required: true
        type: integer
      - description: Настройки публикации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.PublishMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Материал опубликован
          schema:
            $ref: '#/definitions/handlers.PublishMaterialResponse'
        "400":
          description: Неверные параметры запроса
          schema:
            $ref: '#/definitions/handlers.InvalidParametersErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Опубликовать материал
      tags:
      - materials
  /materials/my:
    get:
      consumes:
      - application/json
      description: Возвращает материалы текущего пользователя
      parameters:
      - description: Фильтр по статусу
        enum:
        - draft
        - published
        - archived
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Материалы пользователя
          schema:
            $ref: '#/definitions/models.UserMaterialsResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handlers.InternalErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить материалы пользователя
      tags:
      - materials
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 'Введите: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...